Rewrite all the Ruby in bash
The ruby was just plumbing together a bunch of commands; a task much
better suited for Bash.

Along the way, I brought it up-to-date with modern Heroku runtime, by
changing the output of `bin/release` to be YAML and putting the
FREETDS_DIR environment variable in the `.profile.d` build directory.

Also I added `--strip-components` to the tar command so as to remove the
second instance of the FreeTDS version number configuration option in
the script. This allowed me to add a FREETDS_SOURCE_URL variable to
allow the user to download a newer version of FreeTDS without any source
code changes to the buildpack.